requirements.txt
airtable-python-wrapper==0.15.3
DateTime==4.9
pyairtable==1.3.0
pyTelegramBotAPI==4.7.0
python-dotenv==1.0.1
python-telegram-bot==13.14